arquivos menus unificados | reutilização do arquivo funcoes de forma padrao em vez de varios para cada função
