+++ 06 04/Program.cs
@@ -1,33 +1,33 @@
-﻿using System;
+﻿// using System;
 
 
-namespace aula_06_04
-{
-    class Program
-    {
-        public static void Main(string[] arg)
-        {
-            Console.Clear();
+// namespace aula_06_04
+// {
+//     class Program
+//     {
+//         public static void Main(string[] arg)
+//         {
+//             Console.Clear();
 
-            // exercicio1.podeDirigir();
-            // Console.ReadKey();
-            // exercicio2.definirIdade();
-            // Console.ReadKey();
-            // exercicio3.multiploDeDoisOuTres();
-            // Console.ReadKey();
-            //  exercicio4.definirTemperatura();
-            // Console.ReadKey();
-            // exercicio5.gerarAumento();
-            // Console.ReadKey();
-            // exercicio6.idadeNadador();
-            // Console.ReadKey();
-            // exercicio7.maiorDeTres();
-            // Console.ReadKey();
-            // exercicio8.calcularIMC();
-            // Console.ReadKey();
-            // exercicio9.verificarTriangulo();
-            // Console.ReadKey();
-            // exercicio10.mediaNota();
-        }
-    }
-}
+//             // exercicio1.podeDirigir();
+//             // Console.ReadKey();
+//             // exercicio2.definirIdade();
+//             // Console.ReadKey();
+//             // exercicio3.multiploDeDoisOuTres();
+//             // Console.ReadKey();
+//             //  exercicio4.definirTemperatura();
+//             // Console.ReadKey();
+//             // exercicio5.gerarAumento();
+//             // Console.ReadKey();
+//             // exercicio6.idadeNadador();
+//             // Console.ReadKey();
+//             // exercicio7.maiorDeTres();
+//             // Console.ReadKey();
+//             // exercicio8.calcularIMC();
+//             // Console.ReadKey();
+//             // exercicio9.verificarTriangulo();
+//             // Console.ReadKey();
+//             // exercicio10.mediaNota();
+//         }
+//     }
+// }
