+++ funcoes/funcoes.cs
@@ -1,32 +1,34 @@
 using System;
 
-namespace menuBasicoCsharp.menus
+namespace menuBasicoCsharp.funcoes
 {
     public class Funcoes
     {
-        public void verificarMaior()
+        public static void verificaOMaiorEntre2Numeros()
         {
-            int n1, n2;
-            Console.Clear();
-            Console.Write("Insira um numero: ");
-            int.TryParse(Console.ReadLine(), out n1);
-            Console.Write("Insira ouro: ");
-            int.TryParse(Console.ReadLine(), out n2);
-            if (n1 > n2)
-            {
-                Console.WriteLine($"{n1} é maior que {n2}.");
-            }
-            else if (n2 == n1)
-            {
-                Console.WriteLine("Os numeors são iguais.");
-            }
-            else
-            {
-                Console.WriteLine($"{n2} é maior que {n1}.");
-            }
-            Console.ReadKey();
-        }
 
+            {
+                int n1, n2;
+
+                Console.Clear();
+                n1 = LerInteiro.lerInteiro("Insira um numero: ");
+                n2 = LerInteiro.lerInteiro("Insira outro: ");
+
+                if (n1 > n2)
+                {
+                    Console.WriteLine($"{n1} é maior que {n2}.");
+                }
+                else if (n2 == n1)
+                {
+                    Console.WriteLine("Os numeors são iguais.");
+                }
+                else
+                {
+                    Console.WriteLine($"{n2} é maior que {n1}.");
+                }
+                Console.ReadKey();
+            }
+        }
         public void votar()
         {
             int idade;
